moving some test utils into zipline's test utils file
